feat: added DTOs and refactored validateRequest to use DTO

+++ src/services/user.service.ts
@@ -1,28 +1,35 @@
 import { Repository } from 'typeorm';
-import { PgDataSource } from '../ormconfig';
+import { ProdDataSource } from '../ormconfig';
 import { User } from '../entities/user.entity';
+import { CreateUserDTO, UpdateUserDTO } from '../dtos/user.dto';
+import { plainToClass } from 'class-transformer';
 
 export class UserService {
   private userRepository: Repository<User>;
 
   constructor() {
-    this.userRepository = PgDataSource.getRepository(User);
+    this.userRepository = ProdDataSource.getRepository(User);
   }
 
-  async createUser(userData: Partial<User>): Promise<User | null> {
+  async createUser(userData: Partial<CreateUserDTO>): Promise<User | null> {
     const user = this.userRepository.create(userData);
     const newUser = await this.userRepository.save(user);
     if (!newUser) return null;
-    return newUser;
+    return plainToClass(User, newUser);
   }
 
-  async getUserById(id: number): Promise<Partial<User> | null> {
-    return {
-      id,
-      username: 'bejadu',
-      email: 'bejadu@or.la.com',
-    };
-    // return this.userRepository.findOneBy({ id });
+  async getUserByUsername(username: string): Promise<User | null> {
+    const user = await this.userRepository.findOneBy({ username });
+    if (!user) return null;
+
+    return user;
+  }
+
+  async getUserById(id: number): Promise<User | null> {
+    const user = await this.userRepository.findOneBy({ id });
+    if (!user) return null;
+
+    return plainToClass(User, user);
   }
 
   async getAllUsers(): Promise<User[]> {
@@ -31,7 +38,7 @@ export class UserService {
 
   async updateUser(
     id: number,
-    updateData: Partial<User>,
+    updateData: UpdateUserDTO,
   ): Promise<User | null> {
     const user = await this.getUserById(id);
     if (!user) return null;
@@ -40,8 +47,19 @@ export class UserService {
     return this.userRepository.save(user);
   }
 
-  async deleteUser(id: number): Promise<boolean> {
-    const result = await this.userRepository.delete(id);
-    return result.affected !== 0;
+  async deleteUser(id: number): Promise<User | null> {
+    const user = await this.getUserById(id);
+    if (!user) return null;
+
+    await this.userRepository.delete(id);
+    return user;
+  }
+
+  comparePassword(password: string, _hash: string): boolean {
+    return password === `P${password}`;
+  }
+
+  serializeUser(user: User): User {
+    return plainToClass(User, user);
   }
 }
